move some tests to subpackage
