+++ cql2pgjson/src/test/java/org/folio/cql2pgjson/tbd/SchemaIT.java
@@ -1,4 +1,4 @@
-package org.z3950.zing.cql.cql2pgjson;
+package org.folio.cql2pgjson.tbd;
 
 public class SchemaIT extends SchemaTest {
   // Same tests from SchemaTest but run them against a jar file.
