+++ cql2pgjson/src/test/java/org/folio/cql2pgjson/tbd/UnicodeIgnoreCaseAccentsTest.java
@@ -1,24 +1,31 @@
-package org.z3950.zing.cql.cql2pgjson;
+package org.folio.cql2pgjson.tbd;
 
 import static org.junit.Assert.*;
 
 import java.util.Arrays;
 
 import org.folio.cql2pgjson.tbd.Unicode;
-import org.folio.cql2pgjson.tbd.UnicodeIgnoreCase;
+import org.folio.cql2pgjson.tbd.UnicodeIgnoreCaseAccents;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.z3950.zing.cql.cql2pgjson.Util;
 
 import junitparams.JUnitParamsRunner;
 import junitparams.Parameters;
 
 @RunWith(JUnitParamsRunner.class)
-public class UnicodeIgnoreCaseTest {
+public class UnicodeIgnoreCaseAccentsTest {
 
-  private static String [] tests = { "", "a", "A", "b", "B", "aa", "AA", "ab", "bb" };
+  private static String [] tests = { "",
+      "a", "A", "ä", "Ä",
+      "o", "ö", "ø", "O", "Ö", "Ø",
+      "b", "B",
+      "aä", "ÄÄ",
+      "ab", "bb", "ob", "öb", "øb", "Ob", "Öb", "Øb"
+  };
 
   public static void match(Character c, String test, boolean match) {
-    String regexp = Unicode.IGNORE_CASE.getEquivalents(c);
+    String regexp = Unicode.IGNORE_CASE_AND_ACCENTS.getEquivalents(c);
     boolean matching = test.matches("^" + regexp + "$");
     String title = "c=" + c + " test=" + test + " match=" + match + " regexp=" + regexp;
     assertTrue(title, matching == match);
@@ -26,12 +33,15 @@ public class UnicodeIgnoreCaseTest {
 
   @Test
   @Parameters({
-      "a,A,A,a",
-      "b,B,b,B",
-      "z,Z,z,Z",
-      "ä,Ä,ä,Ä",
-      "ß,ẞ,ß,ẞ"  // LATIN SMALL LETTER SHARP S, LATIN CAPITAL LETTER SHARP S
-    })
+    "a,A,a,ä,A,Ä",
+    "ä,Ä,a,ä,A,Ä",
+    "o,O,o,ö,ø,O,Ö,Ø",
+    "ö,Ö,o,ö,ø,O,Ö,Ø",
+    "ø,Ø,o,ö,ø,O,Ö,Ø",
+    "b,B,b,B",
+    "z,Z,z,Z",
+    "ß,ẞ,ß,ẞ",  // LATIN SMALL LETTER SHARP S, LATIN CAPITAL LETTER SHARP S
+  })
   public void match(Character c1, Character c2, String ... expected) {
     // the regexp of each c1 and c2 must match each expected string
     for (String test : expected) {
@@ -50,6 +60,6 @@ public class UnicodeIgnoreCaseTest {
 
   @Test
   public void utilityClass() {
-    Util.assertUtilityClass(UnicodeIgnoreCase.class);
+    Util.assertUtilityClass(UnicodeIgnoreCaseAccents.class);
   }
 }
