+++ cql2pgjson/src/test/java/org/folio/cql2pgjson/tbd/UnicodeIgnoreCaseTest.java
@@ -1,30 +1,25 @@
-package org.z3950.zing.cql.cql2pgjson;
+package org.folio.cql2pgjson.tbd;
 
 import static org.junit.Assert.*;
 
 import java.util.Arrays;
 
 import org.folio.cql2pgjson.tbd.Unicode;
-import org.folio.cql2pgjson.tbd.UnicodeIgnoreCaseAccents;
+import org.folio.cql2pgjson.tbd.UnicodeIgnoreCase;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.z3950.zing.cql.cql2pgjson.Util;
 
 import junitparams.JUnitParamsRunner;
 import junitparams.Parameters;
 
 @RunWith(JUnitParamsRunner.class)
-public class UnicodeIgnoreCaseAccentsTest {
+public class UnicodeIgnoreCaseTest {
 
-  private static String [] tests = { "",
-      "a", "A", "ä", "Ä",
-      "o", "ö", "ø", "O", "Ö", "Ø",
-      "b", "B",
-      "aä", "ÄÄ",
-      "ab", "bb", "ob", "öb", "øb", "Ob", "Öb", "Øb"
-  };
+  private static String [] tests = { "", "a", "A", "b", "B", "aa", "AA", "ab", "bb" };
 
   public static void match(Character c, String test, boolean match) {
-    String regexp = Unicode.IGNORE_CASE_AND_ACCENTS.getEquivalents(c);
+    String regexp = Unicode.IGNORE_CASE.getEquivalents(c);
     boolean matching = test.matches("^" + regexp + "$");
     String title = "c=" + c + " test=" + test + " match=" + match + " regexp=" + regexp;
     assertTrue(title, matching == match);
@@ -32,15 +27,12 @@ public class UnicodeIgnoreCaseAccentsTest {
 
   @Test
   @Parameters({
-    "a,A,a,ä,A,Ä",
-    "ä,Ä,a,ä,A,Ä",
-    "o,O,o,ö,ø,O,Ö,Ø",
-    "ö,Ö,o,ö,ø,O,Ö,Ø",
-    "ø,Ø,o,ö,ø,O,Ö,Ø",
-    "b,B,b,B",
-    "z,Z,z,Z",
-    "ß,ẞ,ß,ẞ",  // LATIN SMALL LETTER SHARP S, LATIN CAPITAL LETTER SHARP S
-  })
+      "a,A,A,a",
+      "b,B,b,B",
+      "z,Z,z,Z",
+      "ä,Ä,ä,Ä",
+      "ß,ẞ,ß,ẞ"  // LATIN SMALL LETTER SHARP S, LATIN CAPITAL LETTER SHARP S
+    })
   public void match(Character c1, Character c2, String ... expected) {
     // the regexp of each c1 and c2 must match each expected string
     for (String test : expected) {
@@ -59,6 +51,6 @@ public class UnicodeIgnoreCaseAccentsTest {
 
   @Test
   public void utilityClass() {
-    Util.assertUtilityClass(UnicodeIgnoreCaseAccents.class);
+    Util.assertUtilityClass(UnicodeIgnoreCase.class);
   }
 }
