+++ cql2pgjson/src/test/java/org/folio/cql2pgjson/tbd/UnicodeIgnoreNoneTest.java
@@ -1,4 +1,4 @@
-package org.z3950.zing.cql.cql2pgjson;
+package org.folio.cql2pgjson.tbd;
 
 import static org.junit.Assert.*;
